LPS-49142 Rename
