+++ modules/portal/portal-spring-extender/src/META-INF/spring/parent/services-configurator-spring.xml
@@ -8,13 +8,13 @@
 	xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
 	xsi:schemaLocation="http://www.springframework.org/schema/beans http://www.springframework.org/schema/beans/spring-beans.xsd http://www.springframework.org/schema/osgi http://www.springframework.org/schema/osgi/spring-osgi.xsd"
 >
-	<bean id="com.liferay.portal.spring.extender.internal.services.ServicesConfigurator" class="com.liferay.portal.spring.extender.internal.services.impl.ServicesConfiguratorImpl">
+	<bean id="com.liferay.portal.spring.extender.internal.service.ServiceConfigurator" class="com.liferay.portal.spring.extender.internal.service.impl.ServiceConfiguratorImpl">
 		<property name="serviceComponentLocalService" ref="com.liferay.portal.service.ServiceComponentLocalService" />
 	</bean>
-	<bean id="com.liferay.portal.spring.extender.internal.services.listener.ServicesConfiguratorListener" class="com.liferay.portal.spring.extender.internal.services.listener.ServicesConfiguratorListener" abstract="true">
-		<property name="servicesConfigurator" ref="com.liferay.portal.spring.extender.internal.services.ServicesConfigurator" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.listener.ServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.ServiceConfiguratorListener" abstract="true">
+		<property name="serviceConfigurator" ref="com.liferay.portal.spring.extender.internal.service.ServiceConfigurator" />
 	</bean>
-	<bean id="com.liferay.portal.spring.extender.internal.services.listener.StartedServicesConfiguratorListener" class="com.liferay.portal.spring.extender.internal.services.listener.StartedServicesConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.services.listener.ServicesConfiguratorListener" />
-	<bean id="com.liferay.portal.spring.extender.internal.services.listener.StoppedServicesConfiguratorListener" class="com.liferay.portal.spring.extender.internal.services.listener.StoppedServicesConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.services.listener.ServicesConfiguratorListener" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.listener.StartedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.StartedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.listener.ServiceConfiguratorListener" />
+	<bean id="com.liferay.portal.spring.extender.internal.service.listener.StoppedServiceConfiguratorListener" class="com.liferay.portal.spring.extender.internal.service.listener.StoppedServiceConfiguratorListener" parent="com.liferay.portal.spring.extender.internal.service.listener.ServiceConfiguratorListener" />
 	<osgi:reference id="com.liferay.portal.service.ServiceComponentLocalService" availability="mandatory" interface="com.liferay.portal.service.ServiceComponentLocalService" filter="(original.bean=true)" />
 </beans>
